Refactoring/CleanArch.
