+++ EmprestimosLivros.API/EmprestimosLivros.API.csproj
@@ -22,7 +22,11 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="DTOs\" />
+    <Folder Include="Interfaces\" />
+    <Folder Include="Mappings\" />
     <Folder Include="Models\" />
+    <Folder Include="Repositories\" />
   </ItemGroup>
 
   <ItemGroup>
